+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.3.26
+## 1.4.8
 - Chore - upgrade behave from 1.2.6 to 1.3.3
 - Change - reorder core.py imports for BEHAVE_STRIP_STEPS_WITH_TRAILING_COLON
 - Change - from RuntimeError to AssertionError for step failure
@@ -13,6 +13,73 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 - Fix - behave renamed "AssertionError" to "ASSERT FAILED"
 - Fix - behave renamed Status errored to error
 - Fix - consider undefined as failure of test
+
+# 1.4.6
+- Add - keep selenium session alive to prevent browser timeout during long after-scenario hooks
+
+# 1.4.5
+- Change - lazy load report images (except Firefox)
+- Chore - skip image check for default tests
+- Chore - improve some logging
+- Chore - improve some tests
+
+# 1.4.4
+- Fix - colorize debug output in html report
+
+# 1.4.3
+- Change table row count equality from 'equals' to 'exactly'
+
+# 1.4.2
+- Fix - browser logs report
+
+# 1.4.1
+- Change - always all report step logs sections
+- Change - add more logging to fuzzy find
+- Change - preserve newlines in browser log report
+
+# 1.4.0
+- Add - proxy options exposed to end user
+CUCU_PROXY_HOST and CUCU_PROXY_PORT are added as environment varibales that will
+direct the underlying browser to attach to a proxy service. This allows for
+monitoring of traffic between the client and the server during tests.
+
+For example:
+Create HAR files based on captured network traffic.
+Profile network activity from the website under test.
+Compare requested endpoints across .feature files to forecast network traffic.
+Use the exposed variables like so:
+CUCU_PROXY_HOST=localhost CUCU_PROXY_PORT=9009 cucu run features/proxy_test.feature
+
+# 1.3.33
+- Add - enhance fuzzy debug logging
+- Chore - refactor fuzzy.js in prep for enhanced filtering
+
+# 1.3.32
+- Fix - ensure table row matches have same number of cols
+- Add - logging to successful find table
+
+# 1.3.31
+- Add - override cucu_secrets redaction using a config option
+
+# 1.3.30
+- Add - steps to validate tables' row count against an operator
+
+## 1.3.29
+- Fix - option from nth dynamic dropdown
+- Fix - report correct nth number in screenshots
+- Fix - dedup fuzzy find
+- Fix - step for blank nth input
+- Add - screenshot when saw option for dropdown
+- Test - Add - test for nth dynamic dropdown
+
+## 1.3.28
+- Add - lint rule to enforce single space after Feature and Scenario keywords
+
+## 1.3.27
+- Add - step to open a new tab
+
+## 1.3.26
+- Change - Turn on "nth" steps for checking dropdown disablement
 
 ## 1.3.25
 - Add - feature log to html report when available